Remove buld folder from versioning -ignore it

+++ web-client-18LF326-XC8.X/modules/receiver/receiver.h
@@ -11,11 +11,13 @@
 #include <xc.h> // include processor files - each processor file is guarded.  
 #include <stdint.h>
 
+
+
 /* Interfaces Declarations*/
 void receiver_init(void);
 void receiver_push(uint8_t data);
 uint8_t receiver_pop(void);
-uint8_t receiver_getCircBuffFilledDataSize(void);
+uint16_t receiver_getCircBuffFilledDataSize(void);
 uint8_t receiver_task(void);
 void receiver_incrTail(void);
 void receiver_push2FrameBuff(uint8_t data);
